added prop call, moved const to components
