+++ src/ui/Exam/QuestionBankPopUp.tsx
@@ -1,11 +1,12 @@
 import React from 'react'
 import { css } from 'emotion'
-import styled from 'react-emotion'
 import { IQuestionBank, QuestionCount } from 'src/models/QuestionBank/IQuestionBank'
+import {EditButton, Background, Popup, ButtonCool} from 'src/ui/Editor/AnswerEditor/Components'
 
 interface QuestionBankPopUpProps {
     show: boolean
     subjects: IQuestionBank[]
+    onClicked: (id: number[]) => void
 }
 
 interface QuestionBankPopUpState {
@@ -13,64 +14,11 @@ interface QuestionBankPopUpState {
     selected: number[]
 }
 
-const EditButton = styled('button')((props: { color: string }) => ({
-    borderRadius: 5,
-    borderColor: 'green',
-    backgroundColor: 'transparent',
-    padding: 8,
-    fontSize: 14,
-    color: props.color,
-    cursor: 'pointer',
-    textAlign: 'center',
-    width: 438,
-    height: 40,
-}))
-
-const Background = styled('div')({
-    position: 'fixed',
-    top: 0,
-    bottom: 0,
-    left: 0,
-    right: 0,
-    backgroundColor: 'rgba(0,0,0,0.3)',
-    padding: 50,
-});
-
-const Popup = styled('div')({
-    backgroundColor: '#fff',
-    borderRadius: 5,
-    maxWidth: 500,
-    minHeight: 300,
-    margin: '0 auto',
-    padding: 30,
-    position: 'relative',
-});
-
-interface ButonProp {
-    selected: boolean
-}
-
-const ButtonCool = styled('button')((state: ButonProp) => ({
-    borderColor: '#6100ED',
-    borderRadius: 1,
-    backgroundColor: state.selected ? '#E7DFFB' : '#fff',
-    borderTopLeftRadius: 4,
-    borderTopRightRadius: 4,
-    borderBottomLeftRadius: 4,
-    borderBottomRightRadius: 4,
-    color: '#6100ED',
-    textAlign: 'center',
-    width: 438,
-    height: 40,
-    marginBottom: 10,
-    fontSize: 17,
-}));
-
 class QuestionBankPopUp extends React.Component<QuestionBankPopUpProps, QuestionBankPopUpState> {
     public state: QuestionBankPopUpState
     constructor(props) {
         super(props)
-        this.state = ({ show: props.show, selected: [] })
+        this.state = { show: props.show, selected: []}
     }
 
     public hide = () => {
@@ -82,11 +30,9 @@ class QuestionBankPopUp extends React.Component<QuestionBankPopUpProps, Question
         if (!this.isInArray(id, temp)) {
             temp.push(id)
             this.setState({ selected: temp })
-            console.log(this.state.selected)
         } else {
             temp.splice(temp.indexOf(id), 1)
             this.setState({ selected: temp })
-            console.log(this.state.selected)
         }
     }
 
@@ -111,12 +57,13 @@ class QuestionBankPopUp extends React.Component<QuestionBankPopUpProps, Question
     }
 
     private showArray = () => {
-        if (this.state.selected) {
+        if (this.state.selected!=[]) {
             let temp = this.state.selected
             temp.sort()
             this.setState({ selected: temp })
-            console.log(this.state.selected)
+            this.props.onClicked(this.state.selected)
         }
+        this.props.onClicked([])
     }
 
     render() {
